ntproperties/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.5
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.0
django-filter==23.2
Pillow==10.0.1
pydantic==2.4.2
pydantic-core==2.10.1
sqlparse==0.4.4
typing-extensions==4.8.0
